requirements.txt
pyfiglet
rich
psycopg2-binary
python-dotenv